@@ -1,7 +1,7 @@
 ﻿@using Signum.Utilities;
 @using System.IO;
 @using Newtonsoft.Json.Linq;
-@inject Microsoft.AspNetCore.Hosting.IHostingEnvironment hostingEnv
+@inject Microsoft.AspNetCore.Hosting.IWebHostEnvironment hostingEnv
 
 @{
     //FileNotFoundException??? You need to run 'yarn watch'. Requires:
